Reduce number of modules
