+++ assets/data/config/settings.json
@@ -1,20 +1,21 @@
 {
+    ## ENCOUNTER MODULE
     # Adjusts reputation with other factions based on combat engagements
     "starpocalypseEngagementModule": true,
 
+    ## INDUSTRY MODULE
     # Adds stations, patrol hqs, and ground defences as required
     "starpocalypseIndustryModule": true,
 
-    # Damage player fleet (on new game)
-    "starpocalypsePlayerModule": true,
-    # Force player module to run on existing save (on game load) - turn on, load, save, turn off
-    "starpocalypsePlayerModuleForce": false,
+    ## NEW GAME MODULE
+    # Confiscates goodies like nanoforges when starting a new game
+    "starpocalypseNewGameModule": true,
+    # Damage starting fleet, can be enabled without new game module
+    "starpocalypseNewGameModuleDamageStartingFleet": true,
+    # Force new game module to run on existing save (on game load) - turn on, load, save, turn off
+    "starpocalypseNewGameModuleForce": false,
 
-    # Confiscates goodies like nanoforges at procgen time (on new game)
-    "starpocalypseProcgenModule": true,
-    # Force procgen module to run on existing save (on game load) - turn on, load, save, turn off
-    "starpocalypseProcgenModuleForce": false,
-
+    ## SUBMARKET MODULE
     # Damages all ships, removes weapons and combat ships, allows contraband
     "starpocalypseSubmarketModule": true,
     # Disallow black market access with active transponder
