--- src/main/java/starpocalypse/PlayerModule.java
@@ -1,28 +1,0 @@
-package starpocalypse;
-
-import com.fs.starfarer.api.Global;
-import com.fs.starfarer.api.fleet.FleetMemberAPI;
-import java.util.List;
-import lombok.extern.log4j.Log4j;
-import starpocalypse.submarket.ShipDamager;
-
-@Log4j
-public class PlayerModule extends ShipDamager {
-
-    public static void initPlayer(boolean newGame) {
-        boolean hasPlayer = Global.getSettings().getBoolean("starpocalypsePlayerModule");
-        boolean hasPlayerForce = Global.getSettings().getBoolean("starpocalypsePlayerModuleForce");
-        if ((hasPlayer && newGame) || hasPlayerForce) {
-            log.info("Enabling player module");
-            PlayerModule module = new PlayerModule();
-            List<FleetMemberAPI> members = Global.getSector().getPlayerFleet().getFleetData().getMembersListCopy();
-            module.damageShips(members);
-        }
-    }
-
-    private void damageShips(List<FleetMemberAPI> members) {
-        for (FleetMemberAPI member : members) {
-            changeShips(null, null, member);
-        }
-    }
-}
