+++ src/main/java/starpocalypse/industry/NewGameListener.java
@@ -5,9 +5,9 @@ import com.fs.starfarer.api.Global;
 /**
  * One off changes to industries during procedural generation.
  */
-public class ProcgenListener extends IndustryListener {
+public class NewGameListener extends IndustryListener {
 
-    public ProcgenListener() {
+    public NewGameListener() {
         super(new ItemRemover());
     }
